Actual port mappings and brightness table
Note: The collector pullup for the phototransistor has been shrunk from
27kohm to 5.6kohm, and the brightness values expect it to be so. This
way, the collector will be fed enough current to not saturate before a
light intensity of a couple hundred luxes, and the sensor voltage will
remain as a linear function of the light intensity. What on earth was I
thinking when I first figured I'd just use a very high collector
resistance, run it in deep saturation and use a lookup table to
re-linearize the readings?