Do not touch low-side driver state when turning off high side
Low side is driven by NPNs that should instantly turn off and on,
whereas high side is driven by powerful PMOS FETs that take a few CPU
cycles to turn off/on, because the gate capacitance is awfully high. To
ensure all lines receiving current for an identical length of time,
do not close any low-side transistors when high side is at the process
of turning off.